fix: 30.10.23 V3

+++ ewm-service/src/main/java/ru/practicum/category/CategoryServiceImpl.java
@@ -28,8 +28,9 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     public void deleteCategory(long categoryId) {
-        categoryRepository.findById(categoryId)
-                .orElseThrow(() -> new ElementNotFoundException("Категория с ID: " + categoryId + " не найдена"));
+        if (!categoryRepository.existsById(categoryId)) {
+            throw new ElementNotFoundException("Категория с ID: " + categoryId + " не найдена");
+        }
         categoryRepository.deleteById(categoryId);
     }
